gen-ai-course/14_ai_projects/multi_agent_researcher/pyproject.toml
[project]
name = "multi-agent-researcher"
version = "0.1.0"
description = "CrewAI multi-agent research platform with Azure AI backend"
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
    "crewai>=0.30.0",
    "crewai-tools>=0.4.0",
    "langchain>=0.2.0",
    "langchain-openai>=0.1.0",
    "azure-search-documents>=11.4.0",
    "azure-cosmos>=4.6.0",
    "azure-identity>=1.15.0",
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.2.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.4.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
